Make CloningSystem Shared and CloneComps a func
